Set up passes to run in a separate thread
This is a slight re-architecting of the system, although it
required fairly minimal actual code changes.  Every pass will now
have a thread dedicated to it.  This allows the main loop to coninue
following all satellites, and performing tasks such as TLE updates,
prioritization, and so on.

This is needed for issues #3, #7 and #8