Added priority checks when new pass thread starts
There is now a priority field in the sat config file and
a routine that checks it every time a pass starts.  If a new
pass thread starts with a higher priority than the active one,
the active one is told to stop and it switches to the new one.
Otherwise the new one waits until the active pass is complete.

Refs #8.